pyproject.toml
[project]
name = "smokefreelab"
version = "0.1.0"
description = "Bayesian and frequentist A/B testing framework for a simulated smoke-free product e-commerce funnel, built on GA4 sample data."
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [
    { name = "Ahmad Zulfan", email = "azul@example.com" },
]
license = { text = "MIT" }
keywords = [
    "ab-testing",
    "bayesian",
    "ga4",
    "product-analytics",
    "attribution",
    "fmcg",
]

dependencies = [
    # Data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    # BigQuery
    "google-cloud-bigquery>=3.17.0",
    "google-cloud-bigquery-storage>=2.24.0",
    "db-dtypes>=1.2.0",
    # Stats & ML
    "scipy>=1.12.0",
    "statsmodels>=0.14.1",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "shap>=0.44.0",
    # Bayesian
    "pymc>=5.10.0",
    "arviz>=0.17.0",
    # Survival / CLV
    "lifelines>=0.27.8",
    "lifetimes>=0.11.3",
    # Viz
    "plotly>=5.19.0",
    "altair>=5.2.0",
    "matplotlib>=3.8.0",
    # App
    "streamlit>=1.31.0",
    # Config
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    # Utils
    "tqdm>=4.66.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.98.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
    "nbstripout>=0.7.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.29.0",
    "jupyterlab>=4.1.0",
    "pandas-stubs>=2.2.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
attribution = [
    "networkx>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smokefreelab"]

# ────────────────────────────────────────────────────────────────
# Tooling config
# ────────────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",    # pycodestyle, pyflakes
    "I",              # isort
    "N",              # pep8-naming
    "UP",             # pyupgrade
    "B",              # bugbear
    "C4",             # comprehensions
    "SIM",            # simplify
    "RUF",            # ruff-specific
    "D",              # pydocstyle
]
ignore = [
    "D100", "D104",   # missing docstrings in modules/packages (acceptable for now)
    "D203", "D213",   # conflicting docstring rules
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]
"notebooks/**" = ["D", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true

# Overrides: libraries actually imported by in-scope code (src/ + app/ + tests/)
# that don't ship type stubs. Add to this list as new modules come online —
# keep unused entries out so ``warn_unused_configs`` stays green.
[[tool.mypy.overrides]]
module = [
    "arviz.*",
    "lifetimes.*",
    "plotly.*",
    "pymc.*",
    "scipy.*",
    "shap.*",
    "sklearn.*",
    "statsmodels.*",
    "xgboost.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=smokefreelab",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "bigquery: marks tests that hit real BigQuery (opt-in only)",
]

[tool.coverage.run]
source = ["src/smokefreelab"]
branch = true
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 60
